src/python/python_template/src/requirements.txt
boto3==1.34.46
pillow==10.2.0